+++ Resources/js/forms/wifi.js
@@ -42,7 +42,7 @@ $(function () {
                 check_scan();
             }
             function check_scan() {
-                $.get('/scan_wifi?SID=' + SID + '&id=' + random_str(), function (data) {
+                $.get('/scan-wifi?SID=' + SID + '&id=' + random_str(), function (data) {
                     dbg_console.log(data);
                     if (data.pending) {
                         window.setTimeout(check_scan, 1000);